git commit -m ":recycle: refactor: Extração de método - correção de método muito longo"
